Notebook/requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
scipy
safetensors
opencv-python
segment-anything
matplotlib
numpy
Pillow
